+++ dwds/lib/src/debugging/inspector.dart
@@ -15,6 +15,7 @@ import 'package:dwds/src/debugging/libraries.dart';
 import 'package:dwds/src/debugging/location.dart';
 import 'package:dwds/src/debugging/remote_debugger.dart';
 import 'package:dwds/src/loaders/ddc_library_bundle.dart';
+import 'package:dwds/src/loaders/ddc_library_bundle.dart';
 import 'package:dwds/src/readers/asset_reader.dart';
 import 'package:dwds/src/utilities/conversions.dart';
 import 'package:dwds/src/utilities/dart_uri.dart';
@@ -739,7 +740,7 @@ class AppInspector implements AppInspectorInterface {
     return _instanceHelper.metadataHelper.isNativeJsError(instanceRef.classRef);
   }
 
-  /// Request and cache <ScriptRef>s for all the scripts in the application.
+  /// Request and cache `<ScriptRef>`s for all the scripts in the application.
   ///
   /// This populates [_scriptRefsById], [_scriptIdToLibraryId],
   /// [_libraryIdToScriptRefs] and [_serverPathToScriptRef].