+++ dwds/CHANGELOG.md
@@ -1,7 +1,11 @@
-## 24.3.3-wip
+## 24.3.4-wip
+- Added support for some debugging APIs with the DDC library bundle format. - [#2566](https://github.com/dart-lang/webdev/issues/2566), [#2573](https://github.com/dart-lang/webdev/issues/2573)
 
-- Added support for some debugging APIs with the DDC library bundle format. - [#2563](https://github.com/dart-lang/webdev/issues/2563),[#2566](https://github.com/dart-lang/webdev/issues/2566)
+## 24.3.3
+
+- Added support for some debugging APIs with the DDC library bundle format. - [#2563](https://github.com/dart-lang/webdev/issues/2563)
 - Update `DCM` version to `1.26.0-1`
+- Add support for hot restart using the DDC library bundle format.
 
 ## 24.3.2